+++ src/WebHemi/Routing/Result.php
@@ -16,6 +16,63 @@ namespace WebHemi\Routing;
  */
 class Result
 {
-    // Todo implement
-    public $request; // TEST
+    const CODE_FOUND = 200;
+    const CODE_NOT_FOUND = 404;
+    const CODE_BAD_METHOD = 405;
+
+    /** @var int */
+    private $status;
+    /** @var string */
+    private $matchedMiddleware;
+    /** @var array */
+    private $statusReason = [
+        self::CODE_FOUND => 'Resource found.',
+        self::CODE_NOT_FOUND => 'The requested resource cannot be found.',
+        self::CODE_BAD_METHOD => 'Bad request method was used by the client.'
+    ];
+
+    /**
+     * Sets status code.
+     *
+     * @param int $status
+     */
+    public function setStatus($status)
+    {
+        $this->status = $status;
+    }
+
+    /**
+     * Gets status code.
+     *
+     * @return int
+     */
+    public function getStatus()
+    {
+        return $this->status;
+    }
+
+    public function getStatusReason()
+    {
+        return $this->statusReason[$this->status];
+    }
+
+    /**
+     * Sets matched middleware.
+     *
+     * @param string $matchedMiddleware
+     */
+    public function setMatchedMiddleware($matchedMiddleware)
+    {
+        $this->matchedMiddleware = $matchedMiddleware;
+    }
+
+    /**
+     * Gets matched middleware.
+     *
+     * @return string
+     */
+    public function getMatchedMiddleware()
+    {
+        return $this->matchedMiddleware;
+    }
 }